src/OAPM/requirements.txt
click==7.0
pytest-runner==4.4
pytest==4.3.0
requests==2.21.0
kafka-python==1.4.4
